requirements.txt
Django==3.1.1
psycopg2==2.8.7
djangorestframework==3.11.1

requests=2.24.0
apscheduler=3.6.3
django-environ=0.4.5


pytest==6.0.1
pytest-mock==3.3.1
pytest-django==3.9.0
